get_pair_data moved to abstract class
