+++ base_volatility_integration.py
@@ -2,6 +2,8 @@ import logging
 from abc import ABC, abstractmethod
 from typing import Optional, Tuple
 
+from aiohttp import ClientSession
+
 from VolatilityTask.serializers import IntegrationVolatilityMetric
 
 logger = logging.getLogger("asyncio")
@@ -31,6 +33,18 @@ class BaseIntegrationClient(ABC):
         if not interval:
             return None
         return self.interval_mapping.get(interval)
+
+    async def get_pair_data(self, **kwargs) -> list[Optional[list]]:
+        params = await self._prepare_params(**kwargs)
+
+        async with ClientSession() as session:
+            async with session.get(url=self.stock_base_url, params=params) as response:
+                stocks = await response.json()
+                if not response.status == 200:
+                    logger.error(f'{self.name} get pair data error: {stocks["msg"]} {params["symbol"]}')
+                    return []
+                stocks = await response.json()
+                return stocks
 
 
 class StockMarketIntegration(ABC):
